UnitTestFrameworkPkg: Move common includes to their own file
Previously, the UnitTestFrameworkPkgHost.dsc.inc included the entire
UnitTestFrameworkPkgTarget.dsc.inc file. This is unnecessary for
most configurations, so copy the relevant common components to a
separate file.

This is required for stack cookies so that we can have stack
cookies on target based test apps but not on host base test apps.

Signed-off-by: Oliver Smith-Denny <osde@linux.microsoft.com>