BaseTools/CodeQL: Removed unused static function query
This query seems to produce a rate of false positives with some
common patterns in edk2 like passing function pointers for callback.

Due to the usage of `STATIC` instead of `static` particularly for
functions, this query was rarely used in the past. It is removed
here to prevent future false positives.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>